Load an experiment from a model

Loads an experiment from a model specified in a .gaml file and returns an object of class experiment.

load_experiment(experiment, model, dir = "")

Arguments

experiment
The name of the experiment to load.

model
The name of the file from which to load the experiment.

dir
The name of the directory to save the output of the runs for each model. If not specified, name of the gaml file will be used

Details

The rama package contains a library of model in the examples directory of the rama package file hierarchy. These models can be accessed with the system.file() function as explained in the example.

Examples

# Looking at the \code{sir.gaml} file in the \code{examples} directory of the
# \code{rama} library:
# Loading an experiment:
exp1 <- load_experiment("sir", system.file("examples", "sir.gaml", package = "rama"))

#> Creating working directory "" in "/Users/choisy/Dropbox/aaa/r-and-gama/rama/docs/reference".

#> 

#> The directory "sir" is created in the current directory "/Users/choisy/Dropbox/aaa/r-and-gama/rama/docs/reference".

#> 

#> Simulations results will be saved in "/Users/choisy/Dropbox/aaa/r-and-gama/rama/docs/reference/sir".

#> 

#> Loading experiment "sir" from file "sir.gaml"...

#> 

# Checking the class:
class(exp1)

#> [1] "experiment" "data.frame"